fix: restore int cast for cached token holdings count

+++ app/Http/Controllers/Inertia/WalletController.php
@@ -207,7 +207,7 @@ final class WalletController
 
     public function getTokenHoldingsCount(Wallet $wallet): int
     {
-        return Cache::remember('token_holdings_count_'.$wallet->address, now()->addMinutes(10), function () use ($wallet) {
+        return (int) Cache::remember('token_holdings_count_'.$wallet->address, now()->addMinutes(10), function () use ($wallet) {
             return TokenHolder::where('address', $wallet->address)->count();
         });
     }
